PluginStore: Add new plugin type "TTSPlugin"
