+++ client/pluginstore.py
@@ -149,6 +149,7 @@ class PluginStore(object):
         self._translations_dirname = PLUGIN_TRANSLATIONS_DIRNAME
         self._categories_map = {
             'speechhandler': plugin.SpeechHandlerPlugin,
+            'tts': plugin.TTSPlugin,
             'stt': plugin.STTPlugin
         }
